src/requirements.txt
google-api-python-client>=2.0.0
youtube-transcript-api>=0.6.0
yt-dlp>=2023.0.0
youtube-comment-downloader>=0.1.0
pandas>=1.5.0
pathlib2>=2.3.0
python-dotenv>=0.19.0 